racing_line_tools/requirements.txt
numpy<2.0.0
opencv-python<4.10.0
scipy>=1.7.0
scikit-image>=0.19.0
pyyaml>=5.4.0
cvxpy>=1.2.0
matplotlib>=3.5.0

